requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
collection==0.1.6
cycler==0.11.0
cymem==2.0.6
fastai==2.6.3
fastcore==1.4.4
fastdownload==0.0.6
fasterai==0.1.8
fastprogress==1.0.2
fonttools==4.31.2
idna==3.3
imageio==2.16.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
murmurhash==1.0.7
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathy==0.6.1
Pillow==9.0.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.7
pytz==2022.1
PyWavelets==1.3.0
pywin32==302
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.0
scipy==1.8.0
sklearn==0.0
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
thinc==8.0.17
threadpoolctl==3.1.0
tifffile==2022.3.25
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
wincertstore==0.2
xlwt==1.3.0
